fix types

+++ src/xtd.core/include/xtd/diagnostics/stack_frame.h
@@ -29,41 +29,41 @@ namespace xtd {
       stack_frame();
       /// @brief Initializes a new instance of the xtd::diagnostics::stack_frame class that corresponds to a frame above the current stack frame.
       /// @param skip_frame The number of frames up the stack to skip.
-      explicit stack_frame(int32_t skip_frame);
+      explicit stack_frame(size_t skip_frame);
       /// @brief Initializes a new instance of the StackFrame class, optionally capturing source information.
       /// @param need_file_info true to capture the file name, line number, and column number of the stack frame; otherwise, false.
       explicit stack_frame(bool need_file_info);
       /// @brief Initializes a new instance of the StackFrame class that corresponds to a frame above the current stack frame, optionally capturing source information.
       /// @param skip_frame The number of frames up the stack to skip.
       /// @param need_file_info true to capture the file name, line number, and column number of the stack frame; otherwise, false.
-      stack_frame(int32_t skip_frame, bool need_file_info);
+      stack_frame(size_t skip_frame, bool need_file_info);
       /// @brief Initializes a new instance of the StackFrame class that contains only the given file name, and line number.
       /// @param file_name The file name.
       /// @param line_numer The line number in the specified file.
-      stack_frame(const std::string& file_name, size_t line_number);
+      stack_frame(const std::string& file_name, uint32_t line_number);
       /// @brief Initializes a new instance of the StackFrame class that contains only the given file name, line number, and method name.
       /// @param file_name The file name.
       /// @param line_numer The line number in the specified file.
       /// @param method_name The method name.
-      stack_frame(const std::string& file_name, size_t line_number, const std::string& method_name);
+      stack_frame(const std::string& file_name, uint32_t line_number, const std::string& method_name);
       /// @brief Initializes a new instance of the StackFrame class that contains only the given file name, line number, and column number.
       /// @param file_name The file name.
       /// @param line_numer The line number in the specified file.
       /// @param method_name The method name.
       /// @param offset The offset from the code for the method that is being executed.
-      stack_frame(const std::string& file_name, size_t line_number, const std::string& method_name, size_t column_number);
+      stack_frame(const std::string& file_name, uint32_t line_number, const std::string& method_name, uint32_t column_number);
       /// @brief Initializes a new instance of the StackFrame class that contains only the given file name, line number, and column number.
       /// @param file_name The file name.
       /// @param line_numer The line number in the specified file.
       /// @param method_name The method name.
       /// @param column_number The column number in the specified file.
-      stack_frame(const std::string& file_name, size_t line_number, size_t column_number);
+      stack_frame(const std::string& file_name, uint32_t line_number, uint32_t column_number);
       /// @brief Initializes a new instance of the StackFrame class that contains only the given file name, line number, column number, method name, and ofsset.
       /// @param file_name The file name.
       /// @param line_numer The line number in the specified file.
       /// @param method_name The method name.
       /// @param offset The offset from the code for the method that is being executed.
-      stack_frame(const std::string& file_name, size_t line_number, const std::string& method_name, size_t column_number, size_t offset);
+      stack_frame(const std::string& file_name, uint32_t line_number, const std::string& method_name, uint32_t column_number, uint32_t offset);
       /// @cond
       stack_frame(const stack_frame&) = default;
       stack_frame(stack_frame&&) = default;
@@ -92,7 +92,7 @@ namespace xtd {
       /// if (sf.get_offset() != stack_frame::OFFSET_UNKNOWN)
       ///   console::write_line(" Offset: {}", sf.get_offset());
       /// @endcode
-      virtual size_t get_file_column_number() const;
+      virtual uint32_t get_file_column_number() const;
       
       /// @brief Gets the line number in the file that contains the code that is executing. This information is typically extracted from the debugging symbols for the executable.
       /// @return The file line number, or 0 (zero) if the file line number cannot be determined.
@@ -109,7 +109,7 @@ namespace xtd {
       /// if (sf.get_offset() != stack_frame::OFFSET_UNKNOWN)
       ///   console::write_line(" Offset: {}", sf.get_offset());
       /// @endcode
-      virtual size_t get_file_line_number() const;
+      virtual uint32_t get_file_line_number() const;
       
       /// @brief Gets the file name that contains the code that is executing. This information is typically extracted from the debugging symbols for the executable.
       /// @return The file name, or empty ("") if the file name cannot be determined.
@@ -160,7 +160,7 @@ namespace xtd {
       /// if (sf.get_offset() != stack_frame::OFFSET_UNKNOWN)
       ///   console::write_line(" Offset: {}", sf.get_offset());
       /// @endcode
-      virtual size_t get_offset() const;
+      virtual uint32_t get_offset() const;
       
       /// @brief Builds a readable representation of the stack trace.
       /// @return A readable representation of the stack trace.
@@ -171,18 +171,18 @@ namespace xtd {
       /// @endcond
 
       /// @brief Defines the value that is returned from the get_offset() method when the offset is unknown. This field is constant.
-      /// @remarks The value of this constant is std::numric_limit<size_t>::max().
-      static constexpr const size_t OFFSET_UNKNOWN = std::numeric_limits<size_t>::max();
+      /// @remarks The value of this constant is std::numric_limit<uint32_t>::max().
+      static constexpr const uint32_t OFFSET_UNKNOWN = std::numeric_limits<uint32_t>::max();
       
     private:
       friend class stack_trace;
       static std::vector<stack_frame> get_stack_frames(const std::string& str, size_t skip_frames, bool need_file_info);
 
       std::string file_name_;
-      size_t file_line_number_ = 0;
+      uint32_t file_line_number_ = 0;
       std::string method_name_;
-      size_t file_column_number_ = 0;
-      size_t offset_ = OFFSET_UNKNOWN;
+      uint32_t file_column_number_ = 0;
+      uint32_t offset_ = OFFSET_UNKNOWN;
     };
   }
 }
